feat: support fractional shares via BigDecimal quantity fields
Replaces BigInteger with BigDecimal(20,8) for Transaction.quantity and
Position.quantity throughout the domain, services, and tests. Corporate
action math (splits, mergers, spinoffs) now uses MathContext.DECIMAL128
so fractional shares from 3:2 splits (101 → 151.5), exchange ratios
(100 FOX × 0.4517 → 45.17 DIS), and stock dividends are preserved
without integer truncation. Flyway V1.4 migrates the column in MySQL.

+++ src/main/java/com/companyx/equity/model/Position.java
@@ -7,7 +7,6 @@ import lombok.Setter;
 
 import java.io.Serializable;
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.sql.Timestamp;
 
 /**
@@ -22,7 +21,7 @@ public class Position implements Serializable {
         this.user = user;
         this.timestamp = timestamp;
         this.symbol = symbol;
-        this.quantity = BigInteger.ZERO;
+        this.quantity = BigDecimal.ZERO;
         this.value = BigDecimal.ZERO;
         this.realized = BigDecimal.ZERO;
         this.unrealized = BigDecimal.ZERO;
@@ -43,7 +42,7 @@ public class Position implements Serializable {
     private int id;
     private Timestamp timestamp;
     private String symbol;
-    private BigInteger quantity;
+    private BigDecimal quantity;
     private BigDecimal value;
     private BigDecimal realized;
     private BigDecimal unrealized;
